Restructure repository
- Move Go code to repository root
- Move `third_party` code into `internal`
- Update module path to cunicu.li/go-piv
- Remove superfluous LICENSE file
- Remove contribution guidelines (we dont require a CLA in this fork)
