+++ go.mod
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2020 Google LLC
 // SPDX-License-Identifier: Apache-2.0
 
-module github.com/go-piv/piv-go
+module cunicu.li/go-piv
 
 go 1.16
