+++ piv.go
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: 2020 Google LLC
 // SPDX-License-Identifier: Apache-2.0
 
+// Package piv implements management functionality for the YubiKey PIV applet.
 package piv
 
 import (
